flask_backend/requirements.txt
Flask==3.0.2
Flask-Cors==4.0.0
gunicorn==21.2.0
Jinja2==3.1.3
MarkupSafe==2.1.5
marshmallow==3.21.1
numpy==1.26.1
openai==1.14.3
packaging==24.0
pandas==2.1.2
Pillow==9.0.1
pipenv==2023.11.17
platformdirs==4.1.0
protobuf==4.25.3
py4j==0.10.9.7
pyarrow==14.0.0
pyasn1==0.5.0
pycparser==2.22
pydantic==2.6.4
pydantic_core==2.16.3
python-dateutil==2.8.2
python-dotenv==1.0.1
regex==2023.12.25
requests==2.31.0
tiktoken==0.6.0
tqdm==4.66.2
typing-inspect==0.9.0
typing_extensions==4.10.0
tzdata==2023.3
urllib3==2.0.7
validators==0.22.0
virtualenv==20.25.0
Werkzeug==3.0.1
wrapt==1.16.0
yarl==1.9.4
langchain
chromadb